refactor: begin moving query logic from controllers to models
The new `Flight::arriving_at` method can take in an airport code and
select the flights from the database that are arriving at the airport
with that corresponding code. Plans are to include
`Flight::departing_from` and `Flight::departure_time_matches` so the
controller is completely free of query business logic and can just
hand these methods the parameters they need.