feat: create table for flight search results
Relevant data for future booking creation still exists. This uses
HTML5's "form" attribute to make sure the radio button(s) and the
submit button are pointing at the correct information.

The form itself is invisible on the page because its only direct field
is a hidden one for the passenger number that the user selected while
searching for the flight.